:recycle: refactor: refactoring code review
