+++ kok-core/src/main/java/com/kok/kokcore/places/application/usecase/SearchPlaceUseCase.java
@@ -3,7 +3,10 @@ package com.kok.kokcore.places.application.usecase;
 import com.kok.kokcore.places.application.port.in.PlaceInput;
 import com.kok.kokcore.places.domain.model.PlacesResult;
 
-
+/**
+ * 주변 장소 검색 유스케이스
+ * PlaceInput을 받아 검색 결과(PlacesResult)를 반환합니다.
+ */
 public interface SearchPlaceUseCase {
     PlacesResult getPlaces(PlaceInput input) throws Exception;
 }
